Fix all Python lint issues + enforce ruff in CI (#15)
Auditing surfaced 44+ latent Python lint issues across the FastAPI services;
this fixes them and adds a gate so they can't reaccumulate:

- unsorted/deprecated/unused imports across ~20 files (isort + pyupgrade)
- datetime.timezone.utc -> datetime.UTC (7 sites, py3.11+)
- TurnState(str, Enum) -> StrEnum (voice-orchestrator)
- redundant int(round(...)) in mapping-engine to_minor transform
- unused test variable in concierge-agent
- narrowed metering best-effort noqa to the exact rule (S110), documented intent

Adds ruff.toml (F/I/UP/S/RUF ruleset) and a python-lint CI job (uvx ruff check).
All 26 suites green; ruff clean; 132 files headered.

+++ niche-plans/05-abetmigrate/scaffolding/services/mapping-engine/src/mapping.py
@@ -13,8 +13,9 @@ Reliability guarantees from the plan:
 """
 from __future__ import annotations
 
+from collections.abc import Callable
 from dataclasses import dataclass, field
-from typing import Any, Callable
+from typing import Any
 
 # Built-in field transforms available to a mapping rule.
 TRANSFORMS: dict[str, Callable[[Any], Any]] = {
@@ -22,7 +23,7 @@ TRANSFORMS: dict[str, Callable[[Any], Any]] = {
     "lower": lambda v: v.lower() if isinstance(v, str) else v,
     "upper": lambda v: v.upper() if isinstance(v, str) else v,
     "strip": lambda v: v.strip() if isinstance(v, str) else v,
-    "to_minor": lambda v: int(round(float(v) * 100)),  # rupees -> paise
+    "to_minor": lambda v: round(float(v) * 100),  # rupees -> paise (round() returns int)
 }
